distinction between CCatchOutput and PyCatchOutput - initial commit
