+++ obspy/core/tests/test_util_misc.py
@@ -12,7 +12,8 @@ from ctypes import CDLL
 from ctypes.util import find_library
 
 from obspy import UTCDateTime
-from obspy.core.util.misc import CatchOutput, get_window_times
+from obspy.core.util.misc import get_window_times
+from obspy.core.util.capture import PyCatchOutput, CCatchOutput
 
 
 class UtilMiscTestCase(unittest.TestCase):
@@ -23,13 +24,13 @@ class UtilMiscTestCase(unittest.TestCase):
                      platform.python_version_tuple()[0] == "3",
                      "Does not work with Python 3 for some Windows and OSX "
                      "versions")
-    def test_catch_output(self):
+    def test_c_catch_output(self):
         """
         Tests for CatchOutput context manager.
         """
         libc = CDLL(find_library("c"))
 
-        with CatchOutput() as out:
+        with CCatchOutput() as out:
             os.system('echo "abc"')
             libc.printf(b"def\n")
             # This flush is necessary for Python 3, which uses different
@@ -52,11 +53,11 @@ class UtilMiscTestCase(unittest.TestCase):
             self.assertEqual(out.stdout, b"abc\ndef\nghi\njkl\n")
             self.assertEqual(out.stderr, b"123\n456\n")
 
-    def test_catch_output_io(self):
+    def test_py_catch_output_io(self):
         """
-        Tests that CatchOutput context manager does not break I/O.
+        Tests that PyCatchOutput context manager does not break I/O.
         """
-        with CatchOutput():
+        with PyCatchOutput():
             fn = tempfile.TemporaryFile(prefix='obspy')
 
         try:
@@ -65,7 +66,7 @@ class UtilMiscTestCase(unittest.TestCase):
             fn.read(3)
             fn.close()
         except OSError as e:
-            self.fail('CatchOutput has broken file I/O!\n' + str(e))
+            self.fail('PyCatchOutput has broken file I/O!\n' + str(e))
 
     def test_no_obspy_imports(self):
         """
