+++ obspy/core/util/__init__.py
@@ -29,10 +29,10 @@ from obspy.core.util.base import (ALL_MODULES, DEFAULT_MODULES,
                                   NamedTemporaryFile, _read_from_plugin,
                                   create_empty_data_chunk, get_example_file,
                                   get_matplotlib_version, get_script_dir_name)
-from obspy.core.util.misc import (BAND_CODE, CatchOutput, complexify_string,
-                                  guess_delta, loadtxt, score_at_percentile,
-                                  to_int_or_zero)
+from obspy.core.util.misc import (BAND_CODE, complexify_string, guess_delta,
+                                  loadtxt, score_at_percentile, to_int_or_zero)
 from obspy.core.util.obspy_types import (ComplexWithUncertainties, Enum,
                                          FloatWithUncertainties)
 from obspy.core.util.testing import add_doctests, add_unittests
 from obspy.core.util.version import get_git_version as _get_version_string
+from obspy.core.util.capture import PyCatchOutput, CCatchOutput
