+++ obspy/scripts/tests/test_print.py
@@ -8,7 +8,7 @@ import os
 import unittest
 
 from obspy.scripts.print import main as obspy_print
-from obspy.core.util.misc import CatchOutput
+from obspy.core.util.capture import PyCatchOutput
 
 
 class PrintTestCase(unittest.TestCase):
@@ -19,7 +19,7 @@ class PrintTestCase(unittest.TestCase):
                           for x in ['slist.ascii', 'tspair.ascii']]
 
     def test_print(self):
-        with CatchOutput() as out:
+        with PyCatchOutput() as out:
             obspy_print(self.all_files)
 
         self.assertEqual(
@@ -30,7 +30,7 @@ XX.TEST..BHZ | 2008-01-15T00:00:00.025000Z - 2008-01-15T00:00:15.875000Z | 40.0 
         )
 
     def test_print_nomerge(self):
-        with CatchOutput() as out:
+        with PyCatchOutput() as out:
             obspy_print(['--no-merge'] + self.all_files)
 
         self.assertEqual(
